backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.4.1
requests==2.31.0
python-dotenv==1.0.0
aioredis==2.0.1
aiofiles==23.2.0
aiohttp==3.9.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0